Fix modules default export
